+++ src/public/checkers-client/src/components/board/Board.css
@@ -9,36 +9,3 @@
   column-gap: 0;
   background-color: goldenrod;
 }
-
-.red-square {
-  background-color: red;
-}
-
-.black-square {
-  background-color: black
-}
-
-.square {
-  margin: 1px;
-  grid-column: var(--column);
-  grid-row: var(--row);
-}
-
-.piece {
-  border-radius: 50%;
-  width: 78px;
-  height: 78px;
-  border: 1px solid white;
-}
-
-.red-piece {
-  grid-column: var(--column);
-  grid-row: var(--row);
-  background: red;
-}
-
-.black-piece {
-  grid-column: var(--column);
-  grid-row: var(--row);
-  background: black;
-}
